Renamed physics modules to remove solver label
