+++ src/drivers/serac.cpp
@@ -24,7 +24,7 @@
 #include "infrastructure/terminator.hpp"
 #include "mfem.hpp"
 #include "numerics/mesh_utils.hpp"
-#include "physics/nonlinear_solid_solver.hpp"
+#include "physics/nonlinear_solid.hpp"
 #include "serac_config.hpp"
 
 int main(int argc, char* argv[])
