+++ src/physics/nonlinear_solid.hpp
@@ -14,7 +14,7 @@
 #define NONLINSOLID_SOLVER
 
 #include "mfem.hpp"
-#include "physics/base_solver.hpp"
+#include "physics/base_physics.hpp"
 #include "physics/operators/nonlinear_solid_operators.hpp"
 
 namespace serac {
