+++ tests/serac_linelastic_solver.cpp
@@ -10,7 +10,7 @@
 
 #include "mfem.hpp"
 #include "numerics/mesh_utils.hpp"
-#include "physics/elasticity_solver.hpp"
+#include "physics/elasticity.hpp"
 #include "serac_config.hpp"
 
 namespace serac {
