+++ tests/serac_thermal_solver.cpp
@@ -11,7 +11,7 @@
 
 #include "mfem.hpp"
 #include "numerics/mesh_utils.hpp"
-#include "physics/thermal_solver.hpp"
+#include "physics/thermal_conduction.hpp"
 #include "serac_config.hpp"
 
 namespace serac {
